Added less segmented version of Elixir BBHW.
